feat: add documentation refacto: organize requests files feat: add collections requests
