+++ Gemfile
@@ -16,3 +16,5 @@ gem 'rubocop-rake', '~> 0.6.0'
 gem 'rubocop-rspec', '~> 2.11.1'
 
 gem 'debug', '~> 1.5'
+
+gem "yard", "~> 0.9.28"
