+++ lib/kaze_client/request/requests/jobs/cancel_job_request.rb
