+++ lib/kaze_client/request/requests/jobs/update_template_request.rb
